requirement.txt
streamlit
pandas
matplotlib
bcrypt
sqlite3
openai
pillow